Add TVMDBTests
