+++ Sources/TMDBSwift/Models/MovieMDB.swift
@@ -27,10 +27,10 @@ extension MovieMDB {
     }
 
     /// Get the alternative titles for a specific movie id.
-    public class func alternativeTitles(movieID: Int!, country: String? = nil, completion: @escaping (_ clientReturn: ClientReturn, _ altTitles: AlternativeTitlesMDB?) -> Void) {
+    public class func alternativeTitles(movieID: Int!, country: String? = nil, completion: @escaping (_ clientReturn: ClientReturn, _ altTitles: AlternativeMovieTitlesMDB?) -> Void) {
         // language changed to country to avoid modifiying multiple defined functions.
         Client.Movies(String(movieID) + "/alternative_titles", page: nil, language: country) { apiReturn in
-            let data: AlternativeTitlesMDB? = apiReturn.decode()
+            let data: AlternativeMovieTitlesMDB? = apiReturn.decode()
             completion(apiReturn, data)
         }
     }
