+++ Sources/TMDBSwift/Models/TVMDBModel.swift
@@ -9,10 +9,20 @@
 import Foundation
 
 open class ContentRatingsMDB: Decodable {
-  open  var iso_3166_1: String?
+  open var iso_3166_1: String?
   open var rating: String?
 }
 
 open class TVMDB: DiscoverTVMDB {
-  open var genres = [GenresMDB]()
+    open var genres: [GenresMDB]?
+
+    enum CodingKeys: String, CodingKey {
+        case genres
+    }
+
+    public required init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        genres = try? container.decode([GenresMDB]?.self, forKey: .genres)
+        try super.init(from: decoder)
+    }
 }
